requirements.txt
aiomax==0.1.0
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
pydantic==1.10.12
aiohttp==3.8.5
python-dotenv==1.0.0